+++ BarrosoVictor_MaymoDidac_Rockets/src/utilities/ConstantUtilities.java
@@ -1,7 +1,7 @@
 package utilities;
 
 public class ConstantUtilities {
-	public static double maxTime = 5; // time limit of race, and current time
+	public static int maxTime = 5; // time limit of race, and current time
 	public static int length = 30; 		//lenght for race
 	public static double maxAccProplellant = 100; 	
 	public static double fuelTankCapacity = 500; //acc for propellant
